+++ src/environment.h
@@ -11,8 +11,8 @@ public:
         num_actions = actions.size();
     }
     void Render();
-    int Reset();
-    std::tuple<int, int, bool> Step(const std::string& action);
+    size_t Reset();
+    std::tuple<size_t, int, bool> Step(const std::string& action);
     std::vector<std::string> actions;
     int num_states;
     size_t num_actions;
@@ -20,7 +20,7 @@ private:
     int CalculateReward();
     bool CheckTermination();
     std::vector<std::string> states;
-    int current_state = std::distance(states.begin(), std::find(states.begin(), states.end(), "neutral"));
+    size_t current_state = std::distance(states.begin(), std::find(states.begin(), states.end(), "neutral"));
     size_t days_lived = 0;
     size_t days_without_eating = 0;
     size_t max_days = 50;
